requirements.txt

Pillow >= 2.1.0
tweepy==3.8.0
urllib3==1.24.3
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1